test(line-layer): colocate antialiasing render coverage
