+++ test/render/test-cases/line-layer.spec.ts
@@ -2,12 +2,15 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-import {describe} from 'vitest';
-import {runRenderTestSuite} from '../render-test-suite';
+import {describe, expect, test} from 'vitest';
+import {luma} from '@luma.gl/core';
+import {webgl2Adapter} from '@luma.gl/webgl';
+import {runRenderTestSuite, isRenderTestDeviceEnabled} from '../render-test-suite';
 import type {TestCase} from '../deck-test-utils';
+import {measureWebGPUEdges, STROKE_COLOR} from '../webgpu-antialiasing-test-utils';
 
 /* eslint-disable callback-return */
-import {OrthographicView} from '@deck.gl/core';
+import {Deck, OrthographicView} from '@deck.gl/core';
 import {LineLayer} from '@deck.gl/layers';
 import {routes} from 'deck.gl-test/data';
 
@@ -73,13 +76,136 @@ const testCases: TestCase[] = [
   antialiasingTestCase
 ];
 
-describe.each(['webgl', 'webgpu'] as const)('%s', deviceType => {
-  runRenderTestSuite(testCases, deviceType);
+const ANTIALIASING_TEST_WIDTH = 240;
+const ANTIALIASING_TEST_HEIGHT = 180;
+const ANTIALIASING_MEASURE_LINES = [0, 1, 2, 3].map(index => ({
+  sourcePosition: [-110, -70 + index * 42],
+  targetPosition: [110, -64 + index * 51]
+}));
+
+type Coverage = {solid: number; partial: number; levels: number; minimumPartial: number};
+
+async function measureAntialiasingCoverage(antialiasing: boolean): Promise<Coverage> {
+  const container = document.createElement('div');
+  container.style.cssText =
+    `position:absolute;top:0;left:0;width:${ANTIALIASING_TEST_WIDTH}px;` +
+    `height:${ANTIALIASING_TEST_HEIGHT}px;`;
+  document.body.appendChild(container);
+
+  const device = await luma.createDevice({
+    type: 'webgl',
+    adapters: [webgl2Adapter],
+    webgl: {antialias: false},
+    createCanvasContext: {
+      container,
+      width: ANTIALIASING_TEST_WIDTH,
+      height: ANTIALIASING_TEST_HEIGHT,
+      useDevicePixels: false,
+      autoResize: true
+    }
+  });
+
+  const deck = new Deck({
+    device,
+    container,
+    width: ANTIALIASING_TEST_WIDTH,
+    height: ANTIALIASING_TEST_HEIGHT,
+    useDevicePixels: false,
+    views: new OrthographicView(),
+    viewState: {target: [0, 0, 0], zoom: 0}
+  });
+
+  const coverage = await new Promise<Coverage>(resolve => {
+    deck.setProps({
+      layers: [
+        new LineLayer({
+          id: 'line-antialiasing',
+          data: ANTIALIASING_MEASURE_LINES,
+          getColor: [20, 20, 20],
+          getWidth: 2,
+          widthUnits: 'pixels',
+          antialiasing
+        })
+      ],
+      onAfterRender: () => {
+        const gl = (device as any).gl;
+        const pixels = new Uint8Array(ANTIALIASING_TEST_WIDTH * ANTIALIASING_TEST_HEIGHT * 4);
+        gl.readPixels(
+          0,
+          0,
+          ANTIALIASING_TEST_WIDTH,
+          ANTIALIASING_TEST_HEIGHT,
+          gl.RGBA,
+          gl.UNSIGNED_BYTE,
+          pixels
+        );
+        let solid = 0;
+        let partial = 0;
+        let minimumPartial = 255;
+        const levels = new Set<number>();
+        for (let index = 3; index < pixels.length; index += 4) {
+          const alpha = pixels[index];
+          if (alpha === 255) {
+            solid++;
+          } else if (alpha > 0) {
+            partial++;
+            minimumPartial = Math.min(minimumPartial, alpha);
+            levels.add(alpha);
+          }
+        }
+        resolve({solid, partial, levels: levels.size, minimumPartial});
+      }
+    });
+  });
+
+  deck.finalize();
+  device.destroy();
+  container.remove();
+  return coverage;
+}
+
+describe('LineLayer render tests', () => {
+  describe.each(['webgl', 'webgpu'] as const)('%s', deviceType => {
+    runRenderTestSuite(testCases, deviceType);
+  });
+
+  describe('webgl-no-msaa', () => {
+    runRenderTestSuite([antialiasingTestCase], 'webgl', {
+      deviceMode: 'isolated',
+      webgl: {antialias: false}
+    });
+  });
 });
 
-describe('webgl-no-msaa', () => {
-  runRenderTestSuite([antialiasingTestCase], 'webgl', {
-    deviceMode: 'isolated',
-    webgl: {antialias: false}
-  });
+describe.skipIf(!isRenderTestDeviceEnabled('webgl'))('LineLayer#antialiasing', () => {
+  test('adds analytic coverage where the context provides none', async () => {
+    const off = await measureAntialiasingCoverage(false);
+    const on = await measureAntialiasingCoverage(true);
+
+    expect(off.solid, 'lines were drawn').toBeGreaterThan(500);
+    expect(off.partial, 'non-MSAA lines have hard edges by default').toBe(0);
+    expect(on.partial, 'analytic antialiasing feathers the line edges').toBeGreaterThan(300);
+    expect(on.levels, 'coverage is continuous').toBeGreaterThan(40);
+    expect(on.minimumPartial, 'the full centered coverage ramp is rasterized').toBeLessThan(96);
+  }, 60000);
+});
+
+describe.skipIf(!isRenderTestDeviceEnabled('webgpu'))('LineLayer#antialiasing on WebGPU', () => {
+  test('coverage is applied before premultiplication', async () => {
+    const {partial, worstOvershoot} = await measureWebGPUEdges([
+      new LineLayer({
+        id: 'webgpu-line-antialiasing',
+        data: ANTIALIASING_MEASURE_LINES,
+        getColor: STROKE_COLOR,
+        getWidth: 2,
+        widthUnits: 'pixels',
+        antialiasing: true
+      })
+    ]);
+
+    expect(partial, `lines should be feathered (got ${partial} partial pixels)`).toBeGreaterThan(
+      300
+    );
+    expect(worstOvershoot, 'premultiplied red should track alpha').toBeLessThanOrEqual(2);
+  }, 60000);
 });
